+++ src/base/RequestResponse.php
@@ -2,11 +2,9 @@
 
 namespace craft\commerce\omnipay\base;
 
-use Craft;
 use craft\commerce\base\RequestResponseInterface;
 use craft\commerce\models\Transaction;
 use Omnipay\Common\Message\AbstractResponse;
-use Omnipay\Common\Message\ResponseInterface;
 
 /**
  * @author    Pixel & Tonic, Inc. <support@pixelandtonic.com>
@@ -17,12 +15,12 @@ class RequestResponse implements RequestResponseInterface
     /**
      * @var AbstractResponse
      */
-    protected $response;
+    protected AbstractResponse $response;
 
     /**
      * @var Transaction
      */
-    protected $transaction;
+    protected Transaction $transaction;
 
     public function __construct(AbstractResponse $response, Transaction $transaction)
     {
@@ -105,7 +103,7 @@ class RequestResponse implements RequestResponseInterface
     /**
      * @inheritdoc
      */
-    public function redirect()
+    public function redirect(): void
     {
         $this->response->redirect();
     }
@@ -113,10 +111,8 @@ class RequestResponse implements RequestResponseInterface
     /**
      * @inheritdoc
      */
-    public function getData()
+    public function getData(): mixed
     {
         return $this->response->getData();
     }
-
-
 }