+++ src/events/ItemBagEvent.php
@@ -16,10 +16,10 @@ class ItemBagEvent extends Event
     /**
      * @var Order The order
      */
-    public $order;
+    public Order $order;
 
     /**
      * @var mixed The item bag
      */
-    public $items;
+    public mixed $items;
 }